dot-dsl: Moved Edge and Node structs to separate modules
